WIP in loan create and retrive API

+++ db/entity/investment.go
@@ -9,7 +9,7 @@ type Investment struct {
 	LoanID     string `gorm:"type:varchar(255);not null"`
 	InvestorID string `gorm:"type:varchar(255);not null"`
 
-	Amount sql.NullInt16   `gorm:"type:int"`
-	Profit sql.NullInt16   `gorm:"type:int"`
+	Amount sql.NullInt64   `gorm:"type:int"`
+	Profit sql.NullInt64   `gorm:"type:int"`
 	ROI    sql.NullFloat64 `gorm:"type:float"`
 }
